Change project to a java FX application and change open file dialog to not use the windows xp dialog
It is a java fx? I'm not sure, but its allowing me to use the Java FX methods that I needed to open the dialog. Also you will need to use Oracle's JDK 8. AdoptOpenJDK will give an error when you try and create the artifact bvecause it does not have the JavaFX classes. I tried to make refactor everything so that the project structure would be the same as before, but everything broke when i did that :pensive:
