+++ .idea/description.html
@@ -1,0 +1,2 @@
+<html>Simple <b>JavaFX 2.0</b> application that includes simple .fxml file with attached controller and Main class to quick start. Artifact to build JavaFX application is provided.
+</html>
