+++ src/fr/rader/billy/Controller.java
@@ -1,0 +1,4 @@
+package fr.rader.billy;
+
+public class Controller {
+}
