+++ src/fr/rader/billy/gui/inspector/TimelineInspector.java
