+++ src/fr/rader/billy/gui/inspector/listeners/DeleteKeyframesListener.java
