+++ src/fr/rader/billy/gui/inspector/listeners/MovePathListener.java
