+++ src/fr/rader/billy/gui/inspector/listeners/PositionKeyframeListener.java
