+++ src/fr/rader/billy/gui/inspector/listeners/ShiftListener.java
