+++ src/fr/rader/billy/gui/main/MainInterface.form
