+++ src/fr/rader/billy/gui/main/listeners/CopyTimelineListener.java
