+++ src/fr/rader/billy/gui/main/listeners/DeleteTimelineListener.java
