+++ src/fr/rader/billy/gui/main/listeners/EditViewTimelineListener.java
