+++ src/fr/rader/billy/gui/main/listeners/MoveTimelineListener.java
