+++ src/fr/rader/billy/timeline/Interpolator.java
