+++ src/fr/rader/billy/timeline/Keyframe.java
